Replace rofi-blezz with simple sciript
reflect changes in i3 configuration for phryne.

add mbrun config, and put it in install.sh

Change to qutebrowser colors...before it was making qutebrowser appear
to always be in focus.

add rofi_launch_run file that can be used with the rofi_lauunch script
in my scripts repo to behave similarly to rofi-blezz.